feat: move server.py and webui_api.py into the webui/ package with assets in webui/dist/
