+++ .gitignore
@@ -13,7 +13,7 @@
 
 # ==== Build outputs ====
 # WebUI SPA bundled output (built by `mise //admin:build`).
-cafleet/src/cafleet/webui/
+cafleet/src/cafleet/webui/dist/
 
 # Zensical documentation site built by `mise //:docs-build`. Leading slash so
 # the pattern matches the repo-root artifact only, not any future nested `site/`.
